% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/country.R
\name{country_code}
\alias{country_code}
\title{country_code}
\usage{
country_code(origin, input, output)
}
\arguments{
\item{origin}{A vector that contains codes or country names}

\item{input}{Coding type of origin (Enclose in quotes "")}

\item{output}{Coding type returned from origin (Enclose in quotes "")}
}
\description{
country_code
}
\note{
As with the package countrycode the following types are supported: Correlates of War character,
  CoW-numeric, ISO3-character, ISO3-numeric, ISO2-character, IMF numeric, International
  Olympic Committee, FIPS 10-4, FAO numeric, United Nations numeric,
  World Bank character, standardized country names converted by function \code{country_name}.

  The following strings are available respective to the types supported \code{y} or
  \code{z}: "cowc", "cown", "iso3c", "iso3n", "iso2c", "imf",
  "fips104", "fao", "ioc", "un", "wb", "country".
}
\examples{
country_code("Canada", "country", "cown") #returns 20
country_code("Canada", "country", "cowc") #returns "CAN"
}

