% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/country.R
\name{country_name}
\alias{country_name}
\title{country_name}
\usage{
country_name(x)
}
\arguments{
\item{x}{}
}
\value{
A Vector of Country Names
}
\description{
country_name
}
\examples{
country_name("Canada") #returns "Canada"
}

