% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/country.R
\name{country_code}
\alias{country_code}
\title{country_code}
\usage{
country_code(x)
}
\arguments{
\item{x}{}
}
\value{
A Vector of Country Names
}
\description{
country_code
}
\examples{
country_code("Canada") #returns 20
country_code(c("Canada","Bahrain")) #returns c(20, 692)
}

